[1.1.0] Implemented new method for fetching BigTuna's messages Before if BigTuna was "thinking" about an interaction for longer than 500ms then the message would simply be dropped, now the message is refetched every 100ms until a change is found or 5 seconds have passed.
