+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "sega-bass-jr",
-  "version": "1.0.3",
+  "version": "1.1.0",
   "main": "dist/bot.js",
   "scripts": {
     "start": "node .",
