+++ src/lib/functions/evaluate-bait-inventory.ts
@@ -1,6 +1,14 @@
 import { Embed, EmbedBuilder, Message } from 'discord.js';
 import { BaitPriceList, BotColors } from '../constants.js';
-import { capitaliseString } from './capitalise-string.js';
+
+export const capitaliseString = (string: string): string => {
+	const stringList = string.split(' ');
+	const capitalisedString = stringList.map((string) => {
+		return string[0].toUpperCase() + string.slice(1);
+	}).join(' ');
+
+	return capitalisedString;
+};
 
 export const evaluateBaitInventory = (message: Message, baitInventoryEmbed: Embed): void => {
 	const baitList = baitInventoryEmbed.description.split('\n');
@@ -15,7 +23,7 @@ export const evaluateBaitInventory = (message: Message, baitInventoryEmbed: Embe
 	}
 
 	const baitInventoryValueEmbed = new EmbedBuilder()
-		.setDescription(`Your bait inventory is worth approximately **${parseFloat(baitInventoryValue.toFixed(2))}** 🍭!`)
+		.setDescription(`Your bait inventory is worth approximately **${parseFloat((baitInventoryValue).toFixed(2)).toLocaleString()}** 🍭!`)
 		.setColor(BotColors.Positive);
 
 	message.reply({ embeds: [baitInventoryValueEmbed] });
